vidc: vdec: Clearing pmem refernece in port reconfig.
Pmem reference need to be cleared whenever port
reconfig is done. This needs to be done so as to free
up pmem. If there is anyone holding previous reference
we will fail in buffer allocation after port enable.
This issue happens on maximum supported resolutions.
So, in this commit we make sure we release all refernces
before calling allocate buffer again. Same needs to be
made sure from client side.

Change-Id: Iccc8410f3175daf6976e566027dfc0ad5f0f90bd