vidc: venc: Pass user virtual address as key for recon buffers.
With SMMU support added to video driver, it will get a device virtual
address for every userspace address that is registered.
When buffer is de-registered video driver has to unmap the device virtual
address.
User virtual address is used as a key by driver to map with
device virtual address.

Change-Id: Ia5812dcf37d7c6af45e121d08d70e93de1540d2c