vidc: vdec: Fix seek operation for VC1 AP clips
The state machine at frame parser was not detecting start code
in VC1 AP clips when seek operation was performed. This was due
to the mask code used for identifying the start code in source
stream. It was able to identify only 0x00,0x00,0x01,0x0C
0x00,0x00,0x01,0x0D and 0x00,0x00,0x01,0x0E. So modified the mask
code to identify 0x00,0x00,0x01,0x0F start code also. With this
change, state machine is able to resume and detect start code
after the seek operation.

Change-Id: I7b0b78fb1c4a77d6125485bde1879d99f79d138f
CRs-Fixed: 322855