vidc: Remove frame copying logic for MPEG-4/Divx/MPEG-2.
For non-arbitrary bytes' mode and MPEG-4/DivX/MPEG-2, there
is a frame copy logic to make sure codec config data and
first frame data are sent together to the core. But, DivX311
parser sends frame by frame. This is causing problem in buffer
consumption in the core and resulting in video corruption
for the first few frames. Frame copying is not a requirement
from video core. Hence, removed it to fix this issue.

Change-Id: I3d4ccd423eba57f03c4fc1181005e018ede1c26b
CRs-fixed: 327128,326628